1.4: exception logic added

+++ src/main/java/ru/milexe/test_task/service/GroupService.java
@@ -2,10 +2,11 @@ package ru.milexe.test_task.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.milexe.test_task.exception.DataAlreadyExistsException;
+import ru.milexe.test_task.exception.DataNotFoundException;
 import ru.milexe.test_task.entity.GroupEntity;
 import ru.milexe.test_task.entity.StudentEntity;
 import ru.milexe.test_task.model.Group;
-import ru.milexe.test_task.model.Student;
 import ru.milexe.test_task.repository.GroupRepo;
 import ru.milexe.test_task.repository.StudentRepo;
 
@@ -21,11 +22,15 @@ public class GroupService {
     @Autowired
     private StudentRepo studentRepo;
 
-    public Group createGroup(GroupEntity group){
+    public Group createGroup(GroupEntity group) throws DataAlreadyExistsException {
+        if(groupRepo.findById(group.getId()).isPresent() | groupRepo.findByNumber(group.getNumber()) != null)
+            throw new DataAlreadyExistsException("группа уже существует");
         return Group.toModel(groupRepo.save(group));
     }
 
-    public Group getGroup(Long id){
+    public Group getGroup(Long id) throws DataNotFoundException{
+        if(!groupRepo.findById(id).isPresent())
+            throw new DataNotFoundException("группа не найдена");
         return Group.toModel(groupRepo.findById(id).get());
     }
 
@@ -35,18 +40,27 @@ public class GroupService {
         return resultList;
     }
 
-    public Group deleteGroup(Long id){
-        Group deletedGroup = Group.toModel(groupRepo.findById(id).get());
+    public Group deleteGroup(Long id) throws DataNotFoundException, DataAlreadyExistsException {
+        if(!groupRepo.findById(id).isPresent())
+            throw new DataNotFoundException("группа не найдена");
+        GroupEntity entity = groupRepo.findById(id).get();
+        if(!entity.getTimetables().isEmpty())
+            throw new DataAlreadyExistsException("невозможно удалить группу: существует расписание для этой группы");
+        Group deletedGroup = Group.toModel(entity);
         List<StudentEntity> studentsToDelete = studentRepo.findAllByGroup(groupRepo.findById(id));
         studentRepo.deleteAll(studentsToDelete);
         groupRepo.deleteById(id);
         return deletedGroup;
     }
 
-    public Group updateGroup(GroupEntity group){
+    public Group updateGroup(GroupEntity group) throws DataNotFoundException, DataAlreadyExistsException {
+        if(!groupRepo.findById(group.getId()).isPresent())
+            throw new DataNotFoundException("группа не найдена");
         GroupEntity updatedGroup = groupRepo.findById(group.getId()).get();
+        if(groupRepo.findByNumber(group.getNumber()) != null & group.getNumber() != updatedGroup.getNumber())
+            throw new DataAlreadyExistsException("группа уже существует");
         updatedGroup.setNumber(group.getNumber());
         updatedGroup.setSpeciality(group.getSpeciality()); // this logic was added to view list of students in updated group
-        return Group.toModel(groupRepo.save(group));
+        return Group.toModel(groupRepo.save(updatedGroup));
     }
 }
