+++ src/main/java/ru/milexe/test_task/service/StudentService.java
@@ -4,6 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.milexe.test_task.entity.GroupEntity;
 import ru.milexe.test_task.entity.StudentEntity;
+import ru.milexe.test_task.exception.DataAlreadyExistsException;
+import ru.milexe.test_task.exception.DataNotFoundException;
 import ru.milexe.test_task.model.Group;
 import ru.milexe.test_task.model.Student;
 import ru.milexe.test_task.repository.GroupRepo;
@@ -21,7 +23,9 @@ public class StudentService {
     @Autowired
     private GroupRepo groupRepo;
 
-    public Student createStudent(StudentEntity student, Long groupId){
+    public Student createStudent(StudentEntity student, Long groupId) throws DataAlreadyExistsException {
+        if(studentRepo.findById(student.getId()).isPresent())
+            throw new DataAlreadyExistsException("студент уже существует");
         GroupEntity group = groupRepo.findById(groupId).get();
         student.setGroup(group);
         studentRepo.save(student);
@@ -30,7 +34,7 @@ public class StudentService {
 
     public Student getStudent(Long id) throws Exception {
         if(!studentRepo.findById(id).isPresent())
-            throw new Exception("студент не найден");
+            throw new DataNotFoundException("студент не найден");
         return Student.toModel(studentRepo.findById(id).get());
     }
 
@@ -40,13 +44,17 @@ public class StudentService {
         return resultList;
     }
 
-    public Student deleteStudent(Long id){
+    public Student deleteStudent(Long id) throws DataNotFoundException {
+        if(!studentRepo.findById(id).isPresent())
+            throw new DataNotFoundException("студент не найден");
         Student deletedStudent = Student.toModel(studentRepo.findById(id).get());
         studentRepo.deleteById(id);
         return deletedStudent;
     }
 
-    public Student updateStudent(Student student){
+    public Student updateStudent(Student student) throws DataNotFoundException {
+        if(!studentRepo.findById(student.getId()).isPresent())
+            throw new DataNotFoundException("студент не найден");
         StudentEntity updatedStudent = studentRepo.findById(student.getId()).get();
         updatedStudent.setGroup(groupRepo.findByNumber(student.getGroup()));
         updatedStudent.setName(student.getName());
